Extract plugin extraction tests to a separate class
